app/backend/requirements.txt
flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
llama-cpp-python==0.3.7
bokeh===3.4.3
pandas==2.2.3